Removed the testing files and made new dto restructuring to make code more modular
