--- server/src/database/database.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { DatabaseController } from './database.controller';
-
-describe('DatabaseController', () => {
-  let controller: DatabaseController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [DatabaseController],
-    }).compile();
-
-    controller = module.get<DatabaseController>(DatabaseController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
