+++ server/src/friend/dtos/friend.dto.ts
@@ -1,0 +1,8 @@
+export class FriendRequestDto {
+  userId: string;
+  friendId: string;
+  constructor(userId: string, friendId: string) {
+    this.userId = userId;
+    this.friendId = friendId;
+  }
+}
