--- server/src/friend/friend.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { FriendController } from './friend.controller';
-
-describe('FriendController', () => {
-  let controller: FriendController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [FriendController],
-    }).compile();
-
-    controller = module.get<FriendController>(FriendController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
